Add files via upload
Added NonRepeating.js.
Finds the single non-repeating element from the given array when every other element is repeating twice. XOR(^) operator is used to find the single non-repeating element. Since the XOR of the same element twice is 0(x^x =0), and every other element is repeated twice so, if we take the XOR of the given array, all the repeated elements will be canceled, and we will get our desired output.